refactor breakout and benchmarks
