+++ crates/cli/tests/cli_run.rs
@@ -18,7 +18,6 @@ mod cli_run {
     use indoc::indoc;
     use regex::Regex;
     use roc_cli::{CMD_BUILD, CMD_CHECK, CMD_DEV, CMD_FORMAT, CMD_RUN, CMD_TEST};
-    use roc_error_macros::internal_error;
     use roc_reporting::report::strip_colors;
     use roc_test_utils::assert_multiline_str_eq;
     use serial_test::serial;
@@ -373,37 +372,7 @@ mod cli_run {
         let file_name = file_path_from_root(dir_name, roc_filename);
         let mut roc_app_args: Vec<String> = Vec::new();
 
-        // find the workspace directory so we can give roc build script absolute paths
-        let workspace_dir = PathBuf::from(std::env::var("CARGO_MANIFEST_DIR").unwrap());
-
-        // check glue spec is available
-        let zig_glue_path = workspace_dir
-            .join("..")
-            .join("glue")
-            .join("src")
-            .join("ZigGlue.roc");
-        if !zig_glue_path.is_file() {
-            internal_error!("expected ZigGlue.roc at {}", zig_glue_path.display());
-        }
-
-        // check platform folder is available
-        let platform_path = workspace_dir.join("..").join("..").join(dir_name);
-        if !platform_path.is_dir() {
-            internal_error!("expected platform path at {}", platform_path.display());
-        }
-
-        // re-build the platform, expect a build.roc to be next to the test file
-        // set the working directory to the platform folder
-        let build_script_path = std::path::PathBuf::from(&file_name).with_file_name("build.roc");
-        std::process::Command::new("roc")
-            .current_dir(&platform_path)
-            .arg(&build_script_path)
-            .envs(vec![
-                ("ROC", "roc"),
-                ("ZIG_GLUE", zig_glue_path.display().to_string().as_str()),
-            ])
-            .status()
-            .expect(format!("unable to run build script {}", build_script_path.display()).as_str());
+        cli_utils::helpers::rebuild_host(&PathBuf::from(dir_name), &file_name);
 
         for arg in args {
             match arg {
@@ -1023,6 +992,7 @@ mod cli_run {
     // TODO not sure if this cfg should still be here: #[cfg(not(debug_assertions))]
     // this is for testing the benchmarks, to perform proper benchmarks see crates/cli/benches/README.md
     mod test_benchmarks {
+
         #[allow(unused_imports)]
         use super::{TestCliCommands, UseValgrind};
         use cli_utils::helpers::cli_testing_dir;
@@ -1040,6 +1010,8 @@ mod cli_run {
             _use_valgrind: UseValgrind,
         ) {
             let file_name = cli_testing_dir("benchmarks").join(roc_filename);
+
+            cli_utils::helpers::rebuild_host(&cli_testing_dir("benchmarks"), &file_name);
 
             // TODO fix QuicksortApp and then remove this!
             match roc_filename {
