+++ crates/cli_utils/Cargo.toml
@@ -14,6 +14,7 @@ roc_load = { path = "../compiler/load" }
 roc_module = { path = "../compiler/module" }
 roc_reporting = { path = "../reporting" }
 roc_command_utils = { path = "../utils/command" }
+roc_error_macros = { path = "../error_macros" }
 
 bumpalo.workspace = true
 criterion.workspace = true
