+++ examples/breakout/.gitignore
@@ -1,0 +1,3 @@
+platform/*.a
+target/
+build
